Ruby Foundation Work
I worked on setting up the foundations.  There is a feature to allow you
to move "cards" (right now, just numbers) from the deck to a foundation.
It uses modulo to check that it is valid for that foundation, but I
still need to add checks to ensure it is in order.  It could easily be
adapted to take the numeric value of a card object to work it's
calculations.