Foundation & Cards
This version stores the cards in arrays instead of using a database.  It
shuffles the cards before play.  Presently it lists all the cards and
visually displays the top card.  I have images for all the cards
displayed.  The foundations display all cards.  The rules for the
foundations still need tweaked.